remove implicit casts

+++ src/copro-opc5ls.c
@@ -27,7 +27,7 @@ void copro_opc5ls_write(uint16_t addr, uint16_t data) {
    }
 #endif
    if ((addr & 0xFFF8) == 0xFEF8) {
-      tube_parasite_write(addr & 7, data);
+      tube_parasite_write(addr & 7, (uint8_t) data);
       DBG_PRINT("write: %d = %x\r\n", addr & 7, data);
    } else {
       memory[addr] = data;
